Implement upload and remove from google storage

+++ RestaurantApp/Areas/Administrator/Controllers/ProductController.cs
@@ -1,21 +1,25 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Google.Cloud.Storage.V1;
+using Microsoft.AspNetCore.Mvc;
 using RestaurantApp.Core.Contracts;
 using RestaurantApp.Core.Models.Menu;
 
 namespace RestaurantApp.Areas.Administrator.Controllers
 {
-    public class ProductController : BaseAdministratorController
+	public class ProductController : BaseAdministratorController
 	{
 		private readonly IMenuService menuService;
 		private readonly IProductService productService;
+		private readonly StorageClient storageClient;
+		private readonly string bucketName = "bucketstorage123";
 
-        public ProductController(IMenuService menuService, IProductService productService)
-        {
+		public ProductController(IMenuService menuService, IProductService productService, StorageClient storageClient)
+		{
 			this.menuService = menuService;
 			this.productService = productService;
-        }
+			this.storageClient = storageClient;
+		}
 
-        public async Task<IActionResult> Index(string? category)
+		public async Task<IActionResult> Index(string? category)
 		{
 			var model = await menuService.GetMenuAsync(category);
 
@@ -37,37 +41,51 @@ namespace RestaurantApp.Areas.Administrator.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Add(ProductFormModel model, IFormFile imagePath)
 		{
-            if (!ModelState.IsValid)
+			if (!ModelState.IsValid)
 			{
-                model.Categories = await menuService.GetCategoriesAsync();
-                return View(model);
-            }
-
-			if (imagePath != null && imagePath.Length > 0)
+				model.Categories = await menuService.GetCategoriesAsync();
+				return View(model);
+			}
+			var maxFileSize = 2000000;
+			if (imagePath != null && imagePath.Length > 0 && imagePath.Length < maxFileSize)
 			{
 				var fileName = Guid.NewGuid().ToString() + Path.GetExtension(imagePath.FileName);
-				var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/img/menu", fileName);
+				var contentType = imagePath.ContentType;
 
-				using (var fileStream = new FileStream(filePath, FileMode.Create))
+				if (contentType == "image/png" || contentType == "image/jpeg")
 				{
-					await imagePath.CopyToAsync(fileStream);
+					try
+					{
+						var imageObject = await storageClient.UploadObjectAsync(bucketName, fileName, contentType, imagePath.OpenReadStream());
+						model.ImagePath = imageObject.MediaLink;
+					}
+					catch (Exception ex)
+					{
+						TempData["Error"] = "Failed to upload image: " + ex.Message;
+						ModelState.AddModelError("", "Failed to upload image: " + ex.Message);
+						return View(model);
+					}
 				}
-
-				model.ImagePath = "img/menu/" + fileName;
+			}
+			else
+			{
+				TempData["Error"] = "Invalid image format. Only PNG and JPEG are allowed.";
+				ModelState.AddModelError("", "Invalid image format. Only PNG and JPEG are allowed.");
+				return View(model);
 			}
 
 			try
 			{
-                await productService.AddProductAsync(model);
-            }
-            catch (Exception)
+				await productService.AddProductAsync(model);
+			}
+			catch (Exception)
 			{
 
 				return BadRequest();
 			}
 
-            return RedirectToAction(nameof(Index));
-        }
+			return RedirectToAction(nameof(Index));
+		}
 
 		[HttpGet]
 		public async Task<IActionResult> Edit(int id)
@@ -88,23 +106,23 @@ namespace RestaurantApp.Areas.Administrator.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Edit(ProductFormModel model, int id)
 		{
-            if (!ModelState.IsValid)
+			if (!ModelState.IsValid)
 			{
-                model.Categories = await menuService.GetCategoriesAsync();
-                return View(model);
-            }
+				model.Categories = await menuService.GetCategoriesAsync();
+				return View(model);
+			}
 
 			try
 			{
-                await productService.EditProductAsync(model, id);
-            }
-            catch (Exception)
+				await productService.EditProductAsync(model, id);
+			}
+			catch (Exception)
 			{
 				return BadRequest();
 			}
 
-            return RedirectToAction(nameof(Index));
-        }
+			return RedirectToAction(nameof(Index));
+		}
 
 		[HttpPost]
 		public async Task<IActionResult> Remove(int id)
@@ -115,13 +133,26 @@ namespace RestaurantApp.Areas.Administrator.Controllers
 			{
 				var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/", imgPath);
 
-				if (System.IO.File.Exists(filePath))
+				if (filePath != null)
 				{
-					System.IO.File.Delete(filePath);
-				}
-				else
-				{
-					return NotFound("Image not found.");
+					try
+					{
+						if (System.IO.File.Exists(filePath))
+						{
+							System.IO.File.Delete(filePath);
+						}
+						else
+						{
+							var uri = new Uri(imgPath);
+							var imgName = System.Web.HttpUtility.UrlDecode(Path.GetFileName(uri.LocalPath));
+
+							storageClient.DeleteObject(bucketName, imgName);
+						}
+					}
+					catch (Exception)
+					{
+						return RedirectToAction("Error", "Home");
+					}
 				}
 			}
 			else
@@ -131,11 +162,11 @@ namespace RestaurantApp.Areas.Administrator.Controllers
 
 			try
 			{
-                await productService.RemoveProductAsync(id);
-            }
-            catch (Exception)
+				await productService.RemoveProductAsync(id);
+			}
+			catch (Exception)
 			{
-				return BadRequest();
+				return RedirectToAction("Error", "Home");
 			}
 
 			return RedirectToAction(nameof(Index));
